Move public API to root namespace
