+++ src/SqlStreamStore.AcceptanceTests/SqlStreamStoreAcceptanceTests.DeleteEvent.cs
@@ -4,9 +4,8 @@
     using System.Linq;
     using System.Threading.Tasks;
     using Shouldly;
-    using SqlStreamStore.Streams;
     using Xunit;
-    using static Streams.Deleted;
+    using static SqlStreamStore.Deleted;
 
     public partial class StreamStoreAcceptanceTests
     {
