+++ src/SqlStreamStore.AcceptanceTests/SqlStreamStoreAcceptanceTests.ReadAll.cs
@@ -4,7 +4,6 @@
     using System.Linq;
     using System.Threading.Tasks;
     using Shouldly;
-    using SqlStreamStore.Streams;
     using Xunit;
 
     public partial class StreamStoreAcceptanceTests
