+++ src/SqlStreamStore.AcceptanceTests/SqlStreamStoreAcceptanceTests.ReadStream.cs
@@ -5,7 +5,6 @@
     using System.Threading.Tasks;
     using Shouldly;
     using SqlStreamStore.Infrastructure;
-    using SqlStreamStore.Streams;
     using Xunit;
 
     public partial class StreamStoreAcceptanceTests
