+++ src/SqlStreamStore/IStreamStore.cs
@@ -3,7 +3,6 @@
     using System;
     using System.Threading;
     using System.Threading.Tasks;
-    using SqlStreamStore.Streams;
 
     public interface IStreamStore : IReadonlyStreamStore
     {
