+++ src/SqlStreamStore/PageReadStatus.cs
@@ -1,4 +1,4 @@
-namespace SqlStreamStore.Streams
+namespace SqlStreamStore
 {
     public enum PageReadStatus
     {
